разделение файла js на модули
